Modified id type to Long.

+++ src/main/java/com/pika/Onlinegrocerystore/Services/CartService.java
@@ -6,9 +6,9 @@ import com.pika.Onlinegrocerystore.domain.Product;
 import java.util.List;
 
 public interface CartService {
-    Cart addCart(int userId);
-    Cart getCart(int cartID);
-    List<Product> getCartList(int userId);
-    void deleteCartItem(long userId, long itemId);
-    void clearCartItem(long userId);
+    Cart addCart(Long userId);
+    Cart getCart(Long cartID);
+    List<Product> getCartList(Long userId);
+    void deleteCartItem(Long userId, Long itemId);
+    void clearCartItem(Long userId);
 }
