+++ src/main/java/com/pika/Onlinegrocerystore/Services/ManagerService.java
@@ -5,11 +5,11 @@ import com.pika.Onlinegrocerystore.domain.Manager;
 import java.util.List;
 
 public interface ManagerService {
-    Manager findById(int id);
+    Manager findById(Long id);
     List<Manager> findAll();
     void update(Manager customer);
     void create(Manager customer);
-    void deleteById(int id);
+    void deleteById(Long id);
     List<Manager> findByName(String name);
     Manager checkLogin(String name, String password);
 }
